Program 3: Sorting
Compare the execution times of three sorting algorithms using two data sets (one small and one large).  One of the algorithms will use the built-in C++ sort, one algorithm will be of order N2, and the third algorithm will be of order NlogN.  Using identical arrays for each sort and display the sort times in nanoseconds.